search: lift search limits out of repo package into search package
