--- internal/search/constants.go
@@ -1,6 +1,0 @@
-package search
-
-const (
-	DefaultMaxSearchResults          = 30
-	DefaultMaxSearchResultsStreaming = 500
-)
